+++ DryWetMidi/Smf.Interaction/Length/Representations/MusicalLength.cs
@@ -256,14 +256,14 @@ namespace Melanchall.DryWetMidi.Smf.Interaction
                 : LengthUtilities.Subtract(this, length);
         }
 
-        public ILength Multiply(int multiplier)
+        public ILength Multiply(double multiplier)
         {
             ThrowIfArgument.IsNegative(nameof(multiplier), multiplier, "Multiplier is negative.");
 
             return new MusicalLength(Fraction * multiplier);
         }
 
-        public ILength Divide(int divisor)
+        public ILength Divide(double divisor)
         {
             ThrowIfArgument.IsNegative(nameof(divisor), divisor, "Divisor is negative.");
